common function for playwright test
