+++ e2e-tests/common.js
@@ -2,13 +2,12 @@ export const checkLogInOrNot = async(window, expect, userName) => {
   const textVisble = await window.locator('//h1["@aria-label=projects"]', {timeout:3000}).isVisible()
   if (textVisble) {
     const title = await window.textContent('[aria-label=projects]')
-    console.log('app is Logged In')
     await expect(title).toBe('Projects')
 
   } else {
     const welcome = await window.textContent('//*[@id="__next"]/div/div[1]/div/h2')
-    console.log('app is not Logged In')
     await expect(welcome).toBe("Welcome!")
+    await window.reload()
   }
   return textVisble;
 }
@@ -19,3 +18,29 @@ export const filterUser = (json, name) => {
       )
     return filtered
   } 
+
+export const commonJson = async(window, userName, packageInfo, fs) => {
+  const newpath = await window.evaluate(() => Object.assign({}, window.localStorage))
+  const path = require('path');
+  const file = path.join(newpath.userPath, packageInfo.name, 'users', 'users.json');
+  const data = await fs.readFileSync(file);
+  return JSON.parse(data);
+}
+
+export const commonFolder = async (window, userName, packageInfo) => {
+  const newpath = await window.evaluate(() => Object.assign({}, window.localStorage))
+  const path = require('path');
+  return path.join(newpath.userPath, packageInfo.name, 'users', userName.toLowerCase())
+}
+
+export const commonFile = async (window, userName, packageInfo) => {
+  const newpath = await window.evaluate(() => Object.assign({}, window.localStorage))
+  const path = require('path');
+  return path.join(newpath.userPath, packageInfo.name, 'users', 'users.json');
+}
+
+export const removeFolderAndFile = async (fs, folder, userName, window, json, file) => {
+  fs.rmSync(folder, { recursive: true, force: true })
+  const userfilterd = await filterUser(json, userName)
+  await fs.writeFileSync(file, JSON.stringify(userfilterd))
+}
